requirements.txt
numpy==1.24.4
#numba
#scipy
#scikit-learn
#opencv_python
Pillow
torch
torchvision

pytorch
pytorch-lightning
clearml
rich
tqdm
split-folders
simple-parsing
streamlit

